Construct commands using lists, not strings
This is the important one. We are not using a shell to invoke
command-line tools. So adding a string like "--foo --bar" to the list of
arguments to a tool will *not* be the same as adding two flags, `--foo`
and `--bar`. We have to pass each argument as a separate string.